Rename some modules for A005 ruff rule
