+++ scripts/ci/update_starter_projects.py
@@ -11,7 +11,7 @@ from langflow.initial_setup.setup import (
     update_project_file,
     update_projects_components_with_latest_component_versions,
 )
-from langflow.interface.types import get_and_cache_all_types_dict
+from langflow.interface.components import get_and_cache_all_types_dict
 from langflow.services.deps import get_settings_service
 from langflow.services.utils import initialize_services
