+++ src/backend/base/langflow/graph/__init__.py
@@ -1,6 +1,6 @@
 from langflow.graph.edge.base import Edge
 from langflow.graph.graph.base import Graph
 from langflow.graph.vertex.base import Vertex
-from langflow.graph.vertex.types import CustomComponentVertex, InterfaceVertex, StateVertex
+from langflow.graph.vertex.vertex_types import CustomComponentVertex, InterfaceVertex, StateVertex
 
 __all__ = ["CustomComponentVertex", "Edge", "Graph", "InterfaceVertex", "StateVertex", "Vertex"]
