+++ src/backend/tests/performance/test_server_init.py
@@ -58,7 +58,7 @@ async def test_initialize_super_user():
 @pytest.mark.benchmark
 async def test_get_and_cache_all_types_dict():
     """Benchmark get_and_cache_all_types_dict function."""
-    from langflow.interface.types import get_and_cache_all_types_dict
+    from langflow.interface.components import get_and_cache_all_types_dict
 
     settings_service = get_settings_service()
     result = await get_and_cache_all_types_dict(settings_service)
@@ -70,7 +70,7 @@ async def test_get_and_cache_all_types_dict():
 async def test_create_starter_projects():
     """Benchmark creation of starter projects."""
     from langflow.initial_setup.setup import create_or_update_starter_projects
-    from langflow.interface.types import get_and_cache_all_types_dict
+    from langflow.interface.components import get_and_cache_all_types_dict
     from langflow.services.utils import initialize_services
 
     await initialize_services(fix_migration=False)
